refactor(fe): extract wallpaper focal point control
