+++ frontend/core/unit/DashboardThread/Appearance/Wallpaper/TuningPanel/DetailPanel/Gradient/index.tsx
@@ -1,4 +1,4 @@
-import { type KeyboardEvent, type PointerEvent, useEffect, useRef, useState } from 'react'
+import { useEffect, useState } from 'react'
 
 import { COLOR } from '~/const/colors'
 import { cn } from '~/css'
@@ -19,6 +19,7 @@ import useLogic from '../../../useLogic'
 import AngleWheel from '../AngleWheel'
 import GroupItem from '../GroupItem'
 import GroupTitle from '../GroupTitle'
+import FocalPointControl, { type TCenter } from './FocalPointControl'
 import {
   applyGradientSpreadValue,
   buildColorChips,
@@ -30,103 +31,6 @@ import {
 type Props = {
   gradient: TGradientRecipe | null
   canUseAngle: boolean
-}
-
-type TCenter = { x: number; y: number }
-
-const clamp01 = (value: number): number => Math.min(1, Math.max(0, value))
-
-function FocalPointControl({
-  center,
-  label,
-  onChange,
-  onCommit,
-}: {
-  center: TCenter
-  label: string
-  onChange: (center: TCenter) => void
-  onCommit: () => void
-}) {
-  const panelRef = useRef<HTMLButtonElement | null>(null)
-  const [draftCenter, setDraftCenter] = useState(center)
-  const s = useSalon()
-
-  useEffect(() => {
-    setDraftCenter(center)
-  }, [center])
-
-  const updateCenter = (clientX: number, clientY: number): void => {
-    const rect = panelRef.current?.getBoundingClientRect()
-    if (!rect) return
-
-    const nextCenter = {
-      x: clamp01((clientX - rect.left) / rect.width),
-      y: clamp01((clientY - rect.top) / rect.height),
-    }
-
-    setDraftCenter(nextCenter)
-    onChange(nextCenter)
-  }
-
-  const handlePointerDown = (event: PointerEvent<HTMLButtonElement>): void => {
-    event.preventDefault()
-    event.currentTarget.setPointerCapture(event.pointerId)
-    updateCenter(event.clientX, event.clientY)
-  }
-  const handlePointerMove = (event: PointerEvent<HTMLButtonElement>): void => {
-    if (!event.currentTarget.hasPointerCapture(event.pointerId)) return
-    updateCenter(event.clientX, event.clientY)
-  }
-  const handlePointerUp = (event: PointerEvent<HTMLButtonElement>): void => {
-    if (event.currentTarget.hasPointerCapture(event.pointerId)) {
-      updateCenter(event.clientX, event.clientY)
-      event.currentTarget.releasePointerCapture(event.pointerId)
-    }
-    onCommit()
-  }
-  const handleKeyDown = (event: KeyboardEvent<HTMLButtonElement>): void => {
-    const delta = 0.03
-    const nextCenter = {
-      x: draftCenter.x,
-      y: draftCenter.y,
-    }
-
-    if (event.key === 'ArrowLeft') nextCenter.x = clamp01(nextCenter.x - delta)
-    else if (event.key === 'ArrowRight') nextCenter.x = clamp01(nextCenter.x + delta)
-    else if (event.key === 'ArrowUp') nextCenter.y = clamp01(nextCenter.y - delta)
-    else if (event.key === 'ArrowDown') nextCenter.y = clamp01(nextCenter.y + delta)
-    else return
-
-    event.preventDefault()
-    setDraftCenter(nextCenter)
-    onChange(nextCenter)
-    onCommit()
-  }
-
-  return (
-    <button
-      type='button'
-      ref={panelRef}
-      className={s.focalPoint}
-      aria-label={label}
-      onPointerDown={handlePointerDown}
-      onPointerMove={handlePointerMove}
-      onPointerUp={handlePointerUp}
-      onPointerCancel={onCommit}
-      onKeyDown={handleKeyDown}
-    >
-      <span className={s.focalPointVerticalLine} style={{ left: '33.333%' }} />
-      <span className={s.focalPointVerticalLine} style={{ left: '66.667%' }} />
-      <span className={s.focalPointHorizontalLine} style={{ top: '50%' }} />
-      <span
-        className={s.focalPointDot}
-        style={{
-          left: `${draftCenter.x * 100}%`,
-          top: `${draftCenter.y * 100}%`,
-        }}
-      />
-    </button>
-  )
 }
 
 export default function Gradient({ gradient, canUseAngle }: Props) {
